% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_install.R
\name{library_install}
\alias{library_install}
\title{Load a multiple packages, install most recent versions if missing}
\usage{
library_install(pkgs, repos = "http://cran.us.r-project.org", ...)
}
\arguments{
\item{pkgs}{character vector of packages names.}

\item{repos}{character vector, the base URL(s) of the repositories to use,
e.g., the URL of a CRAN mirror such as "https://cloud.r-project.org".}

\item{...}{arguments passed to \code{\link{install.packages}}.}
}
\description{
Load a multiple packages, install most recent versions if missing
}
\examples{
library_install('utils')

\dontrun{
packages <- c('networkD3', 'coreSim')
library_install(packages)
}

}

